Corrected Font
Post Mortem:

A variable weight font requires some specification to be provided. Since it was never provided, the default browser font is used instead, which turns out to be Times New Roman. In my opinion, the font looks rather distinct with proper letter spacing, so it shall be kept as such.